ai/requirements.txt
Flask
python-dotenv
flask-cors
langchain
langchain-openai
